feat: Add Activity model and migration for activity logging
