--- alembic/versions/42bf43768bc8_adding_activity_table.py
@@ -1,172 +1,0 @@
-"""Adding Activity Table
-
-Revision ID: 42bf43768bc8
-Revises:
-Create Date: 2025-03-24 15:23:56.480294
-
-"""
-from typing import Sequence, Union
-
-import sqlalchemy as sa
-
-from alembic import op
-
-# revision identifiers, used by Alembic.
-revision: str = "42bf43768bc8"
-down_revision: Union[str, None] = None
-branch_labels: Union[str, Sequence[str], None] = None
-depends_on: Union[str, Sequence[str], None] = None
-
-
-def upgrade() -> None:
-    """Upgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.create_table(
-        "users",
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("name", sa.String(), nullable=False),
-        sa.Column("email", sa.String(), nullable=False),
-        sa.Column("password", sa.String(), nullable=False),
-        sa.Column("phone", sa.String(), nullable=True),
-        sa.Column("profile_picture_url", sa.String(), nullable=True),
-        sa.Column("created_at", sa.DateTime(), nullable=True),
-        sa.PrimaryKeyConstraint("id"),
-        sa.UniqueConstraint("email"),
-        sa.UniqueConstraint("phone"),
-    )
-    op.create_table(
-        "groups",
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("name", sa.String(), nullable=False),
-        sa.Column(
-            "type", sa.Enum("trip", "home", "other", name="group_type"), nullable=False
-        ),
-        sa.Column("created_by", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["created_by"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "activities",
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=False),
-        sa.Column("timestamp", sa.DateTime(), nullable=True),
-        sa.Column("user_id", sa.UUID(), nullable=False),
-        sa.Column("group_id", sa.UUID(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["group_id"],
-            ["groups.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["user_id"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "expenses",
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("group_id", sa.UUID(), nullable=True),
-        sa.Column("total_amount", sa.Float(), nullable=False),
-        sa.Column("description", sa.Text(), nullable=True),
-        sa.Column(
-            "expense_type",
-            sa.Enum("GROUP", "NON-EXPENSE GROUP", name="expense_type_enum"),
-            nullable=False,
-        ),
-        sa.Column("paid_by", sa.UUID(), nullable=False),
-        sa.Column("created_at", sa.DateTime(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["group_id"],
-            ["groups.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["paid_by"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "group_users",
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("group_id", sa.UUID(), nullable=False),
-        sa.Column("user_id", sa.UUID(), nullable=False),
-        sa.Column("joined_at", sa.DateTime(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["group_id"],
-            ["groups.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["user_id"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "expense_splits",
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("expense_id", sa.UUID(), nullable=False),
-        sa.Column("user_id", sa.UUID(), nullable=False),
-        sa.Column("amount_owed", sa.Float(), nullable=False),
-        sa.Column(
-            "split_type",
-            sa.Enum(
-                "UNEQUALLY",
-                "EQUALLY",
-                "BY SHARES",
-                "BY PERCENTAGE",
-                "BY ADJUSTMENTS",
-                name="split_type_enum",
-            ),
-            nullable=False,
-        ),
-        sa.ForeignKeyConstraint(
-            ["expense_id"],
-            ["expenses.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["user_id"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    op.create_table(
-        "settlements",
-        sa.Column("id", sa.UUID(), nullable=False),
-        sa.Column("expense_split_id", sa.UUID(), nullable=False),
-        sa.Column("payer_id", sa.UUID(), nullable=False),
-        sa.Column("payee_id", sa.UUID(), nullable=False),
-        sa.Column("amount", sa.Float(), nullable=False),
-        sa.Column("is_settled", sa.Boolean(), nullable=True),
-        sa.Column("created_at", sa.DateTime(), nullable=True),
-        sa.ForeignKeyConstraint(
-            ["expense_split_id"],
-            ["expense_splits.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["payee_id"],
-            ["users.id"],
-        ),
-        sa.ForeignKeyConstraint(
-            ["payer_id"],
-            ["users.id"],
-        ),
-        sa.PrimaryKeyConstraint("id"),
-    )
-    # ### end Alembic commands ###
-
-
-def downgrade() -> None:
-    """Downgrade schema."""
-    # ### commands auto generated by Alembic - please adjust! ###
-    op.drop_table("settlements")
-    op.drop_table("expense_splits")
-    op.drop_table("group_users")
-    op.drop_table("expenses")
-    op.drop_table("activities")
-    op.drop_table("groups")
-    op.drop_table("users")
-    # ### end Alembic commands ###
