+++ alembic/versions/7d290cd153e1_adding_activity_table.py
@@ -1,0 +1,98 @@
+"""Adding Activity Table
+
+Revision ID: 7d290cd153e1
+Revises: 23289cbbf519
+Create Date: 2025-03-24 16:47:38.612713
+
+"""
+from typing import Sequence, Union
+
+import sqlalchemy as sa
+from sqlalchemy.dialects import postgresql
+
+from alembic import op
+
+# revision identifiers, used by Alembic.
+revision: str = "7d290cd153e1"
+down_revision: Union[str, None] = "23289cbbf519"
+branch_labels: Union[str, Sequence[str], None] = None
+depends_on: Union[str, Sequence[str], None] = None
+
+
+def upgrade() -> None:
+    """Upgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.create_table(
+        "activities",
+        sa.Column("id", sa.UUID(), nullable=False),
+        sa.Column("description", sa.Text(), nullable=False),
+        sa.Column("timestamp", sa.DateTime(), nullable=True),
+        sa.Column("user_id", sa.UUID(), nullable=False),
+        sa.Column("group_id", sa.UUID(), nullable=True),
+        sa.ForeignKeyConstraint(
+            ["group_id"],
+            ["groups.id"],
+        ),
+        sa.ForeignKeyConstraint(
+            ["user_id"],
+            ["users.id"],
+        ),
+        sa.PrimaryKeyConstraint("id"),
+    )
+    op.alter_column(
+        "expense_splits",
+        "amount_owed",
+        existing_type=sa.NUMERIC(),
+        type_=sa.Float(),
+        existing_nullable=False,
+    )
+    op.alter_column("expenses", "group_id", existing_type=sa.UUID(), nullable=True)
+    op.alter_column(
+        "expenses",
+        "total_amount",
+        existing_type=sa.NUMERIC(),
+        type_=sa.Float(),
+        existing_nullable=False,
+    )
+    op.drop_column("expenses", "date")
+    op.alter_column(
+        "settlements",
+        "amount",
+        existing_type=sa.NUMERIC(),
+        type_=sa.Float(),
+        existing_nullable=False,
+    )
+    # ### end Alembic commands ###
+
+
+def downgrade() -> None:
+    """Downgrade schema."""
+    # ### commands auto generated by Alembic - please adjust! ###
+    op.alter_column(
+        "settlements",
+        "amount",
+        existing_type=sa.Float(),
+        type_=sa.NUMERIC(),
+        existing_nullable=False,
+    )
+    op.add_column(
+        "expenses",
+        sa.Column("date", postgresql.TIMESTAMP(), autoincrement=False, nullable=True),
+    )
+    op.alter_column(
+        "expenses",
+        "total_amount",
+        existing_type=sa.Float(),
+        type_=sa.NUMERIC(),
+        existing_nullable=False,
+    )
+    op.alter_column("expenses", "group_id", existing_type=sa.UUID(), nullable=False)
+    op.alter_column(
+        "expense_splits",
+        "amount_owed",
+        existing_type=sa.Float(),
+        type_=sa.NUMERIC(),
+        existing_nullable=False,
+    )
+    op.drop_table("activities")
+    # ### end Alembic commands ###
